fix: split turn integration tests for 400-line domain test cap (#3876)
Collapsed domain entrypoints exceeded repo.domain_package_test_file_size.
Keep no part-of fragments while splitting each domain into standalone files
under 400 physical lines and extend the integration entrypoint allowlist.
